helm-wikidata.el: Provide actions for each wikidata entity
Previously, there was a single action for storing the functions that
would be executed in those entities. That is, the actions for items,
properties, lexemes and forms were the same.

Now, each entity have a dedicated variable for storing the actions
that need to be performed in each class of entities.